Implement case-insensitive header names
According to the HTTP specification, HTTP headers should be treated as
case-insensitive. To implement this, we retain a mapping from the lower
case name into the name which is specified via serde, and convert header
names to lower case during deserialization, using this mapping to
convert them back if necessary.

The most efficient implementations will be those which don't need any
mappings because their header names are already specified in lower case.
This will avoid a lot of string allocations.